Add support for new style abstractions in cfr
holdem_game is now template class due to having to support multiple types
of abstractions.

Refactored the constructor/generate/write/read functionality in
holdem_abstraction_v2. The class now handles file manipulation and
configuration string parsing itself.


Former-commit-id: 4021f7dceed73465cea1eaa09fff26096cf2e472

+++ src/cfrlib/cfr_solver.ipp
@@ -167,7 +167,7 @@ std::vector<int> cfr_solver<T, U, Data>::get_bucket_counts() const
     std::vector<int> counts;
 
     for (int i = 0; i < ROUNDS; ++i)
-        counts.push_back(abstraction_->get_bucket_count(i));
+        counts.push_back(static_cast<int>(abstraction_->get_bucket_count(i)));
 
     return counts;
 }
